+++ lib/fling_gallery.dart
@@ -2,3 +2,4 @@
 library;
 
 export './src/gallery.dart';
+export './src/layouts/_library.dart';