+++ lib/src/render_object.dart
@@ -12,94 +12,19 @@ class GalleryRenderObject extends RenderBox
     with
         ContainerRenderObjectMixin<RenderBox, GalleryParentData>,
         RenderBoxContainerDefaultsMixin<RenderBox, GalleryParentData> {
-  /// The preferred row height.
-  ///
-  /// The widget will attempt to lay out its children such that the resulting
-  /// row height is as close to this value as possible.
-  double get preferredRowHeight => _preferredRowHeight;
-  double _preferredRowHeight;
-  set preferredRowHeight(double value) {
-    if (_preferredRowHeight == value) {
-      return;
+  GalleryLayoutStrategy get layoutStrategy => _layoutStrategy;
+  GalleryLayoutStrategy _layoutStrategy;
+  set layoutStrategy(GalleryLayoutStrategy newStrategy) {
+    if (newStrategy != _layoutStrategy) {
+      _layoutStrategy = newStrategy;
+      markNeedsLayout();
     }
-    _preferredRowHeight = value;
-    markNeedsLayout();
-  }
-
-  /// The maximum number of items per row.
-  ///
-  /// The widget will force additional items onto new rows, possibly increasing
-  /// the total number of rows. If this is null, there is no maximum, and only
-  /// the items' widths will restrict the items per row.
-  int? get maxRowItems => _maxRowItems;
-  bool get hasMaxRowItems => _maxRowItems != null;
-  int? _maxRowItems;
-  set maxRowItems(int? value) {
-    if (_maxRowItems == value) {
-      return;
-    }
-    _maxRowItems = value;
-    markNeedsLayout();
-  }
-
-  /// The spacing between items in a row.
-  double get horizontalSpacing => _horizontalSpacing;
-  double _horizontalSpacing;
-  set horizontalSpacing(double value) {
-    if (_horizontalSpacing == value) {
-      return;
-    }
-    _horizontalSpacing = value;
-    markNeedsLayout();
-  }
-
-  /// The spacing between rows.
-  double get verticalSpacing => _horizontalSpacing;
-  double _verticalSpacing;
-  set verticalSpacing(double value) {
-    if (_verticalSpacing == value) {
-      return;
-    }
-    _verticalSpacing = value;
-    markNeedsLayout();
-  }
-
-  /// Whether the bottom row should be forced into a full row.
-  bool get forceFill => _forceFill;
-  bool _forceFill;
-  set forceFill(bool value) {
-    if (_forceFill == value) {
-      return;
-    }
-    _forceFill = value;
-    markNeedsLayout();
-  }
-
-  /// The maximum ratio to use when scaling.
-  double get maxScaleRatio => _maxScaleRatio;
-  double _maxScaleRatio;
-  set maxScaleRatio(double value) {
-    if (_maxScaleRatio == value) {
-      return;
-    }
-    _maxScaleRatio = value;
-    markNeedsLayout();
   }
 
   GalleryRenderObject({
+    required GalleryLayoutStrategy layoutStrategy,
     List<RenderBox>? children,
-    required double preferredRowHeight,
-    required int? maxRowItems,
-    required double horizontalSpacing,
-    required double verticalSpacing,
-    required double maxScaleRatio,
-    required bool forceFill,
-  })  : _preferredRowHeight = preferredRowHeight,
-        _maxRowItems = maxRowItems,
-        _horizontalSpacing = horizontalSpacing,
-        _verticalSpacing = verticalSpacing,
-        _maxScaleRatio = maxScaleRatio,
-        _forceFill = forceFill {
+  }) : _layoutStrategy = layoutStrategy {
     addAll(children);
   }
 
@@ -154,7 +79,7 @@ class GalleryRenderObject extends RenderBox
 
     Offset offset = const Offset(0.0, 0.0);
     for (var row in layout.rows) {
-      final rowHeight = row.ratio * preferredRowHeight;
+      final rowHeight = row.ratio * layoutStrategy.preferredRowHeight;
       for (var slot in row.slots) {
         final slotWidth = slot.width * row.ratio;
         final child = slot.child;
@@ -165,9 +90,12 @@ class GalleryRenderObject extends RenderBox
             ),
             parentUsesSize: true);
         (child.parentData! as GalleryParentData).offset = offset;
-        offset = Offset(offset.dx + slotWidth + horizontalSpacing, offset.dy);
+        offset = Offset(
+            offset.dx + slotWidth + layoutStrategy.horizontalSpacing,
+            offset.dy);
       }
-      offset = Offset(0.0, offset.dy + rowHeight + verticalSpacing);
+      offset =
+          Offset(0.0, offset.dy + rowHeight + layoutStrategy.verticalSpacing);
     }
   }
 
@@ -181,12 +109,20 @@ class GalleryRenderObject extends RenderBox
     defaultPaint(context, offset);
   }
 
+  Iterable<RenderBox> get children sync* {
+    var child = firstChild;
+    while (child != null) {
+      yield child;
+      child = childAfter(child);
+    }
+  }
+
   /// Returns an iterator over the sizes of all children, in order.
   Iterable<Size> _childSizes() sync* {
     RenderBox? child = firstChild;
     while (child != null) {
-      final childSize = child
-          .getDryLayout(BoxConstraints.tightFor(height: preferredRowHeight));
+      final childSize = child.getDryLayout(
+          BoxConstraints.tightFor(height: layoutStrategy.preferredRowHeight));
       yield childSize;
       child = childAfter(child);
     }
@@ -194,103 +130,6 @@ class GalleryRenderObject extends RenderBox
 
   /// Calculates the finished layout given parent constraints.
   GalleryLayout _computeRows(BoxConstraints constraints) {
-    final preferredConstraints =
-        BoxConstraints.tightFor(height: preferredRowHeight);
-    final maxRowWidth =
-        constraints.maxWidth.isNaN ? double.infinity : constraints.maxWidth;
-    RenderBox? child = firstChild;
-    double totalWidth = 0.0;
-    double currentRowWidth = 0.0;
-    double ratioWithoutLatestChild = double.infinity;
-    double ratioWithLatestChild = double.infinity;
-    List<GallerySlot> slots = [];
-    List<GalleryRow> rows = [];
-
-    while (child != null) {
-      // The amount of space taken up by the spacing between items
-      final rowWidthSpacingModifier = slots.length * horizontalSpacing;
-      // The amount of space on the row that is useful for laying out children
-      final adjustedMaxRowWidth = maxRowWidth - rowWidthSpacingModifier;
-      // The current child's width, adjusted for the row's height constraints
-      final adjustedChildWidth = child.getDryLayout(preferredConstraints).width;
-
-      currentRowWidth += adjustedChildWidth;
-      ratioWithLatestChild = adjustedMaxRowWidth / currentRowWidth;
-      if (hasMaxRowItems && maxRowItems! <= slots.length) {
-        rows.add(
-          (
-            slots: slots,
-            ratio: ratioWithoutLatestChild,
-          ),
-        );
-        // reset
-        totalWidth = math.max(totalWidth, maxRowWidth);
-        currentRowWidth = 0.0;
-        ratioWithoutLatestChild = double.infinity;
-        slots = [];
-      } else if (currentRowWidth > adjustedMaxRowWidth) {
-        final deviationWithLatestChild = (ratioWithLatestChild - 1.0).abs();
-        final deviationWithoutLatestChild =
-            (ratioWithoutLatestChild - 1.0).abs();
-        double ratio;
-        if (deviationWithoutLatestChild > deviationWithLatestChild) {
-          ratio = ratioWithLatestChild;
-          // keep the new child
-          slots.add((child: child, width: adjustedChildWidth));
-          child = childAfter(child);
-        } else {
-          ratio = ratioWithoutLatestChild;
-        }
-        rows.add(
-          (
-            slots: slots,
-            ratio: ratio,
-          ),
-        );
-        // reset
-        totalWidth = math.max(totalWidth, maxRowWidth);
-        currentRowWidth = 0.0;
-        ratioWithoutLatestChild = double.infinity;
-        slots = [];
-      } else {
-        slots.add((child: child, width: adjustedChildWidth));
-        child = childAfter(child);
-        ratioWithoutLatestChild = adjustedMaxRowWidth / currentRowWidth;
-      }
-    }
-    if (slots.isNotEmpty) {
-      // add the last row
-      if (forceFill || ratioWithLatestChild < maxScaleRatio) {
-        // Scaling up the last row is not so egregious, so do it
-        rows.add(
-          (
-            slots: slots,
-            ratio: ratioWithLatestChild,
-          ),
-        );
-        totalWidth = math.max(totalWidth, maxRowWidth);
-      } else {
-        // Scaling up the last row makes it too big, so default to left aligned
-        final itemsWidth = math.max(totalWidth,
-            slots.fold(0.0, (accumulator, item) => accumulator + item.width));
-        totalWidth = math.max(itemsWidth, constraints.minWidth);
-        rows.add((
-          slots: slots,
-          ratio: totalWidth / itemsWidth,
-        ));
-      }
-    }
-    final totalHeight = math.min(
-        constraints.maxHeight,
-        rows.isNotEmpty
-            ? rows.fold(0.0,
-                    (height, row) => height + row.ratio * preferredRowHeight) +
-                (rows.length - 1) * verticalSpacing
-            : 0.0);
-    return GalleryLayout(
-      rows: rows,
-      width: totalWidth,
-      height: totalHeight.isFinite ? totalHeight : 0.0,
-    );
+    return layoutStrategy.build(this, constraints);
   }
 }